+++ playground/src/app/info_graph.rs
@@ -10,6 +10,9 @@ use dot_ix::{
 };
 use leptos::*;
 
+#[cfg(target_arch = "wasm32")]
+use super::QUERY_PARAM_DIAGRAM_ONLY;
+
 const INFO_GRAPH_DEMO: &str = include_str!("info_graph_example.yaml");
 
 /// User provided info graph source.
@@ -23,47 +26,67 @@ const QUERY_PARAM_SRC: &str = "src";
 #[cfg(target_arch = "wasm32")]
 fn info_graph_src_init(set_info_graph_src: WriteSignal<String>) {
     use lz_str::decompress_from_encoded_uri_component;
-    use web_sys::{Document, Url};
+    use web_sys::{console, Document, Url, UrlSearchParams};
+    use js_sys::Array;
 
     create_effect(move |_| {
         if let Some(window) = web_sys::window() {
-            let url_search_params = Url::new(&String::from(window.location().to_string()))
-                .expect("Expected URL to be valid.")
-                .search_params();
-            let info_graph_src_initial = url_search_params
-                .get(QUERY_PARAM_SRC)
-                .map(|src| {
-                    if src.contains("\n") {
-                        // Treat src as plain yaml
-                        src
-                    } else {
-                        // Try deserialize/serialize src as lz_str
-                        decompress_from_encoded_uri_component(&src).map_or_else(
-                            || format!("# deserialize src error: invalid data"),
-                            |s| {
-                                String::from_utf16(&s).unwrap_or_else(|_| {
-                                    format!("# deserialize src error: invalid data")
-                                })
-                            },
-                        )
+            let url_search_params = {
+                let url = Url::new(&String::from(window.location().to_string()))
+                    .expect("Expected URL to be valid.");
+
+                let hash = url.hash();
+                if hash.is_empty() {
+                    Some(url.search_params())
+                } else {
+                    match UrlSearchParams::new_with_str(hash.as_str()) {
+                        Ok(search_params) => Some(search_params),
+                        Err(error) => {
+                            let message = Array::new_with_length(1);
+                            message.set(0, error);
+                            console::log(&message);
+                            None
+                        }
                     }
-                })
-                .unwrap_or_else(|| String::from(INFO_GRAPH_DEMO));
+                }
+            };
 
-            set_info_graph_src.set(info_graph_src_initial);
+            if let Some(url_search_params) = url_search_params {
+                let info_graph_src_initial = url_search_params
+                    .get(QUERY_PARAM_SRC)
+                    .map(|src| {
+                        if src.contains("\n") {
+                            // Treat src as plain yaml
+                            src
+                        } else {
+                            // Try deserialize/serialize src as lz_str
+                            decompress_from_encoded_uri_component(&src).map_or_else(
+                                || format!("# deserialize src error: invalid data"),
+                                |s| {
+                                    String::from_utf16(&s).unwrap_or_else(|_| {
+                                        format!("# deserialize src error: invalid data")
+                                    })
+                                },
+                            )
+                        }
+                    })
+                    .unwrap_or_else(|| String::from(INFO_GRAPH_DEMO));
 
-            // Hack: Get Tailwind CSS from CDN to reevaluate document.
-            set_timeout(
-                move || {
-                    let _ = window
-                        .document()
-                        .as_ref()
-                        .and_then(Document::body)
-                        .as_deref()
-                        .map(|element| element.append_with_str_1(""));
-                },
-                Duration::from_millis(200),
-            );
+                set_info_graph_src.set(info_graph_src_initial);
+
+                // Hack: Get Tailwind CSS from CDN to reevaluate document.
+                set_timeout(
+                    move || {
+                        let _ = window
+                            .document()
+                            .as_ref()
+                            .and_then(Document::body)
+                            .as_deref()
+                            .map(|element| element.append_with_str_1(""));
+                    },
+                    Duration::from_millis(200),
+                );
+            }
         } else {
             set_info_graph_src.set(String::from("# Could not extract search params."));
         }
@@ -171,10 +194,29 @@ pub fn InfoGraph(diagram_only: ReadSignal<bool>) -> impl IntoView {
                     let src_compressed = compress_to_encoded_uri_component(&info_graph_src);
                     if let Some(window) = web_sys::window() {
                         let url = {
-                            let u = web_sys::Url::new(&String::from(window.location().to_string()))
+                            let url = web_sys::Url::new(&String::from(window.location().to_string()))
                                 .expect("Expected URL to be valid.");
-                            u.search_params().set(QUERY_PARAM_SRC, &src_compressed);
-                            u.to_string()
+
+                            // Remove this in a few versions.
+                            url.search_params().delete(QUERY_PARAM_SRC);
+                            url.search_params().delete(QUERY_PARAM_DIAGRAM_ONLY);
+
+                            let fragment = {
+                                let mut fragment = String::with_capacity(QUERY_PARAM_SRC.len() + src_compressed.len() + 64);
+                                fragment.push_str(QUERY_PARAM_SRC);
+                                fragment.push_str("=");
+                                fragment.push_str(&src_compressed);
+
+                                if diagram_only.get() {
+                                    fragment.push_str("&");
+                                    fragment.push_str(QUERY_PARAM_DIAGRAM_ONLY);
+                                    fragment.push_str("=true");
+                                }
+
+                                fragment
+                            };
+                            url.set_hash(&fragment);
+                            url.to_string()
                                 .as_string()
                                 .expect("# Failed to decode src parameter")
                         };